backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.9.0
websockets==13.0
openpyxl==3.1.5
aiofiles==24.1.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2